ADDED: Facebook login now works and inflates the login page first
