+++ Beakon/app/src/main/java/erica/beakon/LoginPage.java
@@ -1,7 +1,66 @@
 package erica.beakon;
 
-import com.facebook.FacebookSdk;
-public class LoginPage extends  {
+import android.app.Activity;
+import android.content.DialogInterface;
+import android.content.Intent;
+import android.os.Bundle;
+import android.util.Log;
+import android.view.View;
+import android.widget.Button;
 
+import com.facebook.CallbackManager;
+import com.facebook.CustomTabActivity;
+import com.facebook.FacebookActivity;
+import com.facebook.FacebookCallback;
+import com.facebook.FacebookException;
+import com.facebook.FacebookSdk;
+import com.facebook.login.LoginManager;
+import com.facebook.login.LoginResult;
+import com.facebook.login.widget.LoginButton;
+
+public class LoginPage extends Activity {
+    private CallbackManager callbackManager;
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        FacebookSdk.sdkInitialize(getApplicationContext());
+
+        callbackManager = CallbackManager.Factory.create();
+
+
+
+
+        setContentView(R.layout.login_fragment);
+
+
+        LoginButton loginButton = (LoginButton) findViewById(R.id.login_button);
+        loginButton.setReadPermissions("email");
+
+        final Intent intent = new Intent(this, MainActivity.class);
+
+        loginButton.registerCallback(callbackManager, new FacebookCallback<LoginResult>() {
+            @Override
+            public void onSuccess(LoginResult loginResult) {
+                Log.d("FacebookLogin", "onSuccess" + loginResult);
+                startActivity(intent);
+            }
+
+            @Override
+            public void onCancel() {
+                Log.d("FacebookLogin", "Login attempt cancelled");
+            }
+
+            @Override
+            public void onError(FacebookException e) {
+                Log.d("FacebookLogin", e.getMessage());
+            }
+        });
+    }
+
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+        callbackManager.onActivityResult(requestCode, resultCode, data);
+    }
 
 }
